requirements.txt
addict==2.4.0
matplotlib==3.2.2
numpy==1.22.3
opencv_python_headless==4.5.4.60
pandas==1.4.3
Pillow==9.2.0
ptflops==0.6.7
PyYAML==6.0
requests==2.22.0
scipy==1.5.0
seaborn==0.11.2
setuptools==45.2.0
thop==0.0.31.post2005241907
timm==0.5.4
torch==1.11.0+cu113
torchvision==0.12.0+cu113
tqdm==4.50.0
wandb==0.12.9
yacs==0.1.8
